requirements.txt
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Flask==1.0.2
Jinja2==2.10.1
MarkupSafe==1.1.1
WTForms==2.2.1
Werkzeug==0.15.3
gunicorn==20.0.4
itsdangerous==0.24
oauthlib==2.1.0
passlib[argon2]
psycopg2==2.8.4
requests-oauthlib==0.8.0
requests==2.23.0
